[add] 계정삭제 라우터에 요청한 seller정보와 입력된 이메일 비교 추가

+++ libs/nest-modules/src/lib/seller/seller.controller.ts
@@ -7,6 +7,7 @@ import {
   Patch,
   Post,
   Query,
+  UnauthorizedException,
   UseGuards,
   ValidationPipe,
 } from '@nestjs/common';
@@ -20,6 +21,8 @@ import {
 import { JwtAuthGuard } from '../_nest-units/guards/jwt-auth.guard';
 import { MailVerificationService } from '../auth/mailVerification.service';
 import { SellerService } from './seller.service';
+import { SellerInfo } from '../_nest-units/decorators/sellerInfo.decorator';
+import { UserPayload } from '../auth/auth.interface';
 
 @Controller('seller')
 export class SellerController {
@@ -57,8 +60,15 @@ export class SellerController {
   }
 
   // 판매자 계정 삭제
+  @UseGuards(JwtAuthGuard)
   @Delete()
-  public async deleteSeller(@Body('email') email: string) {
+  public async deleteSeller(
+    @Body('email') email: string,
+    @SellerInfo() sellerInfo: UserPayload,
+  ) {
+    if (email !== sellerInfo.sub) {
+      throw new UnauthorizedException('본인의 계정이 아니면 삭제할 수 없습니다.');
+    }
     return this.sellerService.deleteOne(email);
   }
